zwesta-v2-professional/backend/requirements.txt
# FastAPI Backend Requirements
# Zwesta Trading System v2

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.11.0
cryptography==41.0.7

# Trading Integrations
MetaTrader5==5.0.45
python-binance==1.0.17

# Notifications
twilio==8.10.0
python-dotenv==1.0.0
aiohttp==3.9.1

# Reports & Files
reportlab==4.0.7
weasyprint==59.3
openpyxl==3.11.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0
asyncio==3.4.3
pandas==2.1.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
python-json-logger==2.0.7
